Functions assignment Start!
Moved the conditionals assignments into the old assignments folder.
created renamed copies of old folders in the main repository. Since
this is the final week for this class I am going to go all out and
create a full website around this assignment. but first i am going to
make sure the assignment is ready to turn in without all that.